[t] fake client moved to independent package
